backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
motor
pydantic[email]
passlib[bcrypt]
PyJWT[crypto]
aiofiles
aiosmtplib
razorpay
beanie
python-multipart
httpx